src/requirements.txt
streamlit
pandas
python-dotenv
langchain
#google-generativeai
groq    